Move gcrane cobra stuff under cmd/gcrane/cmd
This mirrors the crane structure.
