+++ cmd/gcrane/cmd/root.go
@@ -10,7 +10,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gcrane
+package cmd
 
 import (
 	"os"
